Add files via upload
As part of our computer vision course, we were asked to recognize objects using the SIFT descriptor. Our application will take an image as input and output the name of the object contained in the image. We will test and compare our program on two different image bases
to compare the accuracy of these bases